通过userid查询用户信息新增查询用户所有角色

+++ ZR.Model/System/SysUser.cs
@@ -99,7 +99,7 @@ namespace ZR.Model.System
         /// </summary>
         [SugarColumn(IsIgnore = true)]
         [EpplusIgnore]
-        public int[] RoleIds { get; set; }
+        public long[] RoleIds { get; set; }
         /// <summary>
         /// 岗位集合
         /// </summary>
